+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 This file contains highlights of what changes on each version of the [Reverse-Proxy.js](https://www.npmjs.com/package/@cedx/reverse-proxy) package.
 
-# Version 3.2.0
+# Version 4.0.0
+- Breaking change: removed the `Application#log` method.
 - The response object is also provided in `request` events.
 - The routing table of a `Server` is now exposed as the `routes` property.
+- Using the [Morgan](https://www.npmjs.com/package/morgan) module to handle the logging.
 
 ## Version 3.1.0
 - Added support for the [Node Security Platform](https://nodesecurity.io) reports.